+++ peppy/project.py
@@ -134,6 +134,7 @@ class Project(PathExAttMap):
         self._samples = []
         self[SAMPLE_EDIT_FLAG_KEY] = False
         self.is_private = False
+        self.progressbar = False
 
         # table indexes can be specified in config or passed to the object constructor
         # That's the priority order:
@@ -473,7 +474,8 @@ class Project(PathExAttMap):
             for s in track(
                 self.samples,
                 description="Removing sample attributes",
-                disable=not self.is_sample_table_large,
+                disable=not all([self.is_sample_table_large,
+                                 self.progressbar]),
             ):
                 for attr in to_remove:
                     _del_if_in(s, attr)
@@ -490,7 +492,8 @@ class Project(PathExAttMap):
             for s in track(
                 self.samples,
                 description="Applying constant sample attributes",
-                disable=not self.is_sample_table_large,
+                disable=not all([self.is_sample_table_large,
+                                 self.progressbar]),
             ):
                 for attr, val in to_append.items():
                     if attr not in s:
@@ -506,7 +509,8 @@ class Project(PathExAttMap):
             for sample in track(
                 self.samples,
                 description="Applying synonymous sample attributes",
-                disable=not self.is_sample_table_large,
+                disable=not all([self.is_sample_table_large,
+                                 self.progressbar]),
             ):
                 for attr, new in synonyms.items():
                     if attr in sample:
@@ -636,7 +640,8 @@ class Project(PathExAttMap):
                 for sample_id in track(
                     sample_names_list,
                     description="Detecting duplicate sample names",
-                    disable=not self.is_sample_table_large,
+                    disable=not all([self.is_sample_table_large,
+                                     self.progressbar]),
                 )
                 if sample_names_list.count(sample_id) > 1
             ]
@@ -677,7 +682,8 @@ class Project(PathExAttMap):
             for sample in track(
                 self.samples,
                 description=f"Merging subsamples, adding sample attrs: {', '.join(subsample_table.keys())}",
-                disable=not self.is_sample_table_large,
+                disable=not all([self.is_sample_table_large,
+                                 self.progressbar]),
             ):
                 sample_colname = self.st_index
                 if sample_colname not in subsample_table.columns:
@@ -768,7 +774,8 @@ class Project(PathExAttMap):
         for sample in track(
             self.samples,
             description="Implying sample attributes",
-            disable=not self.is_sample_table_large,
+            disable=not all([self.is_sample_table_large,
+                             self.progressbar]),
         ):
             for implication in implications:
                 implier_attrs = list(implication[IMPLIED_IF_KEY].keys())
@@ -817,7 +824,8 @@ class Project(PathExAttMap):
         for sample in track(
             self.samples,
             description="Deriving sample attributes",
-            disable=not self.is_sample_table_large,
+            disable=not all([self.is_sample_table_large,
+                             self.progressbar]),
         ):
             for attr in derivations:
                 if not hasattr(sample, attr):